requirements.txt
matplotlib
numpy
pillow
opencv-contrib-python
pandas
scikit-learn
joblib
seaborn